+++ src/test-fixtures/create-wrapped-component.js
